{{cookiecutter.project_slug}}/src/apps/a12n/api/throttling.py
from rest_framework.throttling import AnonRateThrottle

from core.api.throttling import ConfigurableThrottlingMixin


class AuthAnonRateThrottle(ConfigurableThrottlingMixin, AnonRateThrottle):
    """Throttle for any authorization views."""

    scope = "anon-auth"
